+++ app/metrics-adapters/sentry.ts
@@ -27,7 +27,7 @@ export default class Sentry extends BaseAdapter {
 
   // eslint-disable-next-line ember/classic-decorator-hooks
   init(): void {
-    if (ENV.environment === 'test') {
+    if (ENV.env === 'test') {
       this.enabled = false;
     }
   }